pyproject.toml
[project]
name = "image-deteriorator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "albumentationsx>=2.1.3",
    "click>=8.3.2",
    "opencv-python>=4.13.0.92",
    "tqdm>=4.67.3",
]

[tool.uv]
package = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes (errors)
    "UP",  # pyupgrade (modernize code for newer Python)
    "B",   # flake8-bugbear (common bugs and security issues)
    "SIM", # flake8-simplify (simplify code)
    "I",   # isort (import sorting),
    "W",
    "C901"
]
ignore = []

